Implement Initial Authorization For Test

+++ Api/src/Infrastructure/DatingApp.Infrastructure.Persistence/Repositories/GenericRepository.cs
@@ -17,7 +17,7 @@ namespace DatingApp.Infrastructure.Persistence.Repositories
         {
             _dbContext = dbContext;
         }
-        public async Task<T> GetByIdAsync(long id)
+        public async Task<T> GetByIdAsync(int id)
         {
             return await _dbContext.Set<T>().FindAsync(id);
         }
